+++ src/crypto/crypto_aes.h
@@ -6,7 +6,6 @@
 #include "crypto/crypto_cipher.h"
 #include "crypto/crypto_keys.h"
 #include "crypto/crypto_util.h"
-#include "allocated_buffer.h"
 #include "env.h"
 #include "v8.h"
