+++ src/crypto/crypto_bio.cc
@@ -22,7 +22,6 @@
 #include "crypto/crypto_bio.h"
 #include "base_object-inl.h"
 #include "memory_tracker-inl.h"
-#include "allocated_buffer-inl.h"
 #include "util-inl.h"
 
 #include <openssl/bio.h>
