+++ src/crypto/crypto_common.cc
@@ -1,4 +1,3 @@
-#include "allocated_buffer-inl.h"
 #include "base_object-inl.h"
 #include "env-inl.h"
 #include "node_buffer.h"
