+++ src/crypto/crypto_dh.cc
@@ -1,9 +1,8 @@
 #include "crypto/crypto_dh.h"
-#include "crypto/crypto_keys.h"
-#include "crypto/crypto_groups.h"
-#include "allocated_buffer-inl.h"
 #include "async_wrap-inl.h"
 #include "base_object-inl.h"
+#include "crypto/crypto_groups.h"
+#include "crypto/crypto_keys.h"
 #include "env-inl.h"
 #include "memory_tracker-inl.h"
 #include "threadpoolwork-inl.h"
