+++ src/crypto/crypto_ec.cc
@@ -1,9 +1,8 @@
 #include "crypto/crypto_ec.h"
-#include "crypto/crypto_common.h"
-#include "crypto/crypto_util.h"
-#include "allocated_buffer-inl.h"
 #include "async_wrap-inl.h"
 #include "base_object-inl.h"
+#include "crypto/crypto_common.h"
+#include "crypto/crypto_util.h"
 #include "env-inl.h"
 #include "memory_tracker-inl.h"
 #include "node_buffer.h"
