+++ src/crypto/crypto_hash.cc
@@ -1,5 +1,4 @@
 #include "crypto/crypto_hash.h"
-#include "allocated_buffer-inl.h"
 #include "async_wrap-inl.h"
 #include "base_object-inl.h"
 #include "env-inl.h"
