+++ src/crypto/crypto_hkdf.cc
@@ -1,8 +1,7 @@
 #include "crypto/crypto_hkdf.h"
-#include "crypto/crypto_keys.h"
-#include "allocated_buffer-inl.h"
 #include "async_wrap-inl.h"
 #include "base_object-inl.h"
+#include "crypto/crypto_keys.h"
 #include "env-inl.h"
 #include "memory_tracker-inl.h"
 #include "threadpoolwork-inl.h"
