+++ src/crypto/crypto_rsa.h
@@ -7,7 +7,6 @@
 #include "crypto/crypto_keygen.h"
 #include "crypto/crypto_keys.h"
 #include "crypto/crypto_util.h"
-#include "allocated_buffer.h"
 #include "env.h"
 #include "memory_tracker.h"
 #include "v8.h"
