+++ src/crypto/crypto_scrypt.cc
@@ -1,7 +1,6 @@
 #include "crypto/crypto_scrypt.h"
-#include "crypto/crypto_util.h"
-#include "allocated_buffer-inl.h"
 #include "async_wrap-inl.h"
+#include "crypto/crypto_util.h"
 #include "env-inl.h"
 #include "memory_tracker-inl.h"
 #include "node_buffer.h"
