+++ src/env-inl.h
@@ -25,11 +25,11 @@
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
 #include "aliased_buffer.h"
-#include "allocated_buffer-inl.h"
 #include "callback_queue-inl.h"
 #include "env.h"
 #include "node.h"
 #include "node_context_data.h"
+#include "node_internals.h"
 #include "node_perf_common.h"
 #include "util-inl.h"
 #include "uv.h"
@@ -42,6 +42,16 @@
 #include <utility>
 
 namespace node {
+
+NoArrayBufferZeroFillScope::NoArrayBufferZeroFillScope(
+    IsolateData* isolate_data)
+    : node_allocator_(isolate_data->node_allocator()) {
+  if (node_allocator_ != nullptr) node_allocator_->zero_fill_field()[0] = 0;
+}
+
+NoArrayBufferZeroFillScope::~NoArrayBufferZeroFillScope() {
+  if (node_allocator_ != nullptr) node_allocator_->zero_fill_field()[0] = 1;
+}
 
 inline v8::Isolate* IsolateData::isolate() const {
   return isolate_;
@@ -979,7 +989,7 @@ inline uv_buf_t Environment::allocate_managed_buffer(
   std::unique_ptr<v8::BackingStore> bs =
       v8::ArrayBuffer::NewBackingStore(isolate(), suggested_size);
   uv_buf_t buf = uv_buf_init(static_cast<char*>(bs->Data()), bs->ByteLength());
-  released_allocated_buffers()->emplace(buf.base, std::move(bs));
+  released_allocated_buffers_.emplace(buf.base, std::move(bs));
   return buf;
 }
 
@@ -987,18 +997,12 @@ inline std::unique_ptr<v8::BackingStore> Environment::release_managed_buffer(
     const uv_buf_t& buf) {
   std::unique_ptr<v8::BackingStore> bs;
   if (buf.base != nullptr) {
-    auto map = released_allocated_buffers();
-    auto it = map->find(buf.base);
-    CHECK_NE(it, map->end());
+    auto it = released_allocated_buffers_.find(buf.base);
+    CHECK_NE(it, released_allocated_buffers_.end());
     bs = std::move(it->second);
-    map->erase(it);
+    released_allocated_buffers_.erase(it);
   }
   return bs;
-}
-
-std::unordered_map<char*, std::unique_ptr<v8::BackingStore>>*
-    Environment::released_allocated_buffers() {
-  return &released_allocated_buffers_;
 }
 
 inline void Environment::ThrowError(const char* errmsg) {
